refactor: move azuread_authentication_strength_policy to correct service package
