+++ internal/services/conditionalaccess/registration.go
@@ -3,7 +3,9 @@
 
 package conditionalaccess
 
-import "github.com/hashicorp/terraform-provider-azuread/internal/tf/pluginsdk"
+import (
+	"github.com/hashicorp/terraform-provider-azuread/internal/tf/pluginsdk"
+)
 
 type Registration struct{}
 
@@ -29,8 +31,7 @@ func (r Registration) SupportedDataSources() map[string]*pluginsdk.Resource {
 // SupportedResources returns the supported Resources supported by this Service
 func (r Registration) SupportedResources() map[string]*pluginsdk.Resource {
 	return map[string]*pluginsdk.Resource{
-		"azuread_authentication_strength_policy": authenticationStrengthPolicyResource(),
-		"azuread_named_location":                 namedLocationResource(),
-		"azuread_conditional_access_policy":      conditionalAccessPolicyResource(),
+		"azuread_named_location":            namedLocationResource(),
+		"azuread_conditional_access_policy": conditionalAccessPolicyResource(),
 	}
 }
